fix route seperation and naming convention
