+++ routes/users-route.js
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { getUsers } from "../controllers/users-controller.js";
+
+const router = Router();
+
+router.get("/", getUsers);
+
+export default router;
